requirements.txt
requests>=2.20,<=3.0
selenium==3.141.0
pandas==1.0.5
beautifulsoup4==4.9.1
python_dateutil==2.8.1
bs4==0.0.1
lxml==4.5.2
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
webdriver-manager==3.2.2
pep8==1.7.1
autopep8==1.5.4
ipython==7.16.1
spotipy>=2.16,<=3.0
fuzzywuzzy==0.18.0
statsmodels==0.11.1
transliterate==1.10.2
unicode==2.7
unicodecsv==0.14.1
yandex-music==0.1.2
